+++ lib/utp.dart
@@ -1,6 +1,4 @@
-/// Support for doing something awesome.
-///
-/// More dartdocs go here.
+/// UTP library , include [ServerUTPSocket],[UTPSocketClient],[UTPSocket]
 library utp;
 
 export 'src/utp_protocol_implement.dart';